requirements.txt
numpy
scipy
matplotlib
pytest
Cython
fastapi
uvicorn
httpx
jinja2
weasyprint
requests

pywebview
